ecommerce-backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
stripe==7.7.0
python-decouple==3.8
Pillow==10.1.0
psycopg2-binary==2.9.9
django-extensions==3.2.3
setuptools==80.9.0
